+++ prosebot-editor/api/rename_file.php
@@ -7,7 +7,11 @@
  * RESPONSE: string Confirmation or error
  */
 
-header("Access-Control-Allow-Origin: *");
+require_once(__DIR__.'/vendor/autoload.php');
+$dotenv = Dotenv\Dotenv::createImmutable(__DIR__);
+$dotenv->load();
+
+header("Access-Control-Allow-Origin: ".$_ENV['CLIENT']);
 header("Content-Type: application/json; charset=UTF-8");
 header("Access-Control-Allow-Methods: PUT");
 header("Access-Control-Allow-Headers: Content-Type, Access-Control-Allow-Headers, Authorization, X-Requested-With");
@@ -15,7 +19,7 @@ header("Access-Control-Allow-Headers: Content-Type, Access-Control-Allow-Headers
 $data = json_decode(file_get_contents('php://input'));
 if (isset($_GET["name"]) && isset($_GET["context"]) && isset($_GET["lang"])) {
     $dir = __DIR__.'/../templates/'.$_GET["context"].'/'.$_GET["lang"].'/';
-    if (file_exists($dir.$_GET["name"]) && property_exists($data, "filename")) {
+    if (file_exists($dir.$_GET["name"]) && $data !== null && property_exists($data, "filename")) {
         rename($dir.$_GET["name"], $dir.$data->filename);
         echo json_encode(array("message" => "File name updated."));
         http_response_code(200);
@@ -24,6 +28,7 @@ if (isset($_GET["name"]) && isset($_GET["context"]) && isset($_GET["lang"])) {
         echo json_encode(array("message" => "File does not exist."));
         http_response_code(400);
     }
+    else http_response_code(200);
 }
 else {
     echo json_encode(array("message" => "Bad request."));
